use this information to write information on the airdrop:
our fund aims to fully decentralize the ownership and governance of the fund. token holders will earn weekly fees in USDC on their chain of their choice. which is how we revolutionize the fund model and give back to the people who make the fund valuable. you. the people. to be eligble, all u need to do is to own LP tokens of the investment token of your choice. more tokens invested, mean bigger airdrop, means bigger rewards for the holders. don't miss out the oppurtunity to own the future!